Move EventQueue to own file
